requirements.txt
streamlit
pandas
python-dotenv
litellm
pypdf
python-docx
tiktoken
numpy

# Tests (optionnel)
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0